src/requirements.txt
asgiref==3.5.1
certifi==2022.5.18
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
Markdown==3.3.7
mysql-connector-python==8.0.28
Pillow==9.0.1
protobuf==3.20.1
PyJWT==2.4.0
python-environ==0.4.54
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
